+++ packages/styleguide/.storybook/components/PropsTable/Elements.tsx
@@ -1,6 +1,5 @@
-import { styled } from '@storybook/theming';
-import { boxShadow, spacing } from '@codecademy/gamut-styles';
-import { css } from '@emotion/react';
+import styled from '@emotion/styled';
+import { themed, variant } from '@codecademy/gamut-styles/src';
 
 export const PropGroupTooltip = styled.ul`
   display: none;
@@ -8,73 +7,56 @@ export const PropGroupTooltip = styled.ul`
   bottom: calc(100% + 10px);
   left: 50%;
   transform: translate(-50%, 0);
-  padding: ${spacing[8]} ${spacing[12]};
-  background-color: ${({ theme }) => theme.color.light};
-  border: 1px solid ${({ theme }) => theme.color.medium};
+  padding: ${themed('spacing.8')} ${themed('spacing.12')};
+  border: 1px solid ${themed('colors.navy')};
+  background-color: ${themed('colors.white')};
   list-style: none;
   margin: 0;
-  border-radius: 6px;
-  ${boxShadow(1)}
+  z-index: 5;
 `;
 
 export const PropItem = styled.li`
-  margin: ${spacing[8]} 0;
+  margin: ${themed('spacing.8')} 0;
   font-size: 11px;
-  font-family: ${({ theme }) => theme.typography.fonts.mono};
+  font-family: ${themed('fontFamily.accent')};
   text-align: center;
   text-transform: none;
-  color: ${({ theme }) => theme.color.dark};
+  color: ${themed('colors.gray-700')};
 `;
 
-export const HeaderColumn = styled.div``;
+const tagVariants = variant({
+  normal: {
+    textColor: 'navy',
+    borderColor: 'navy',
+  },
+  selected: {
+    textColor: 'white',
+    backgroundColor: 'navy',
+    borderColor: 'navy',
+  },
+});
 
-export const PropGroupTag = styled.span<{ active?: boolean }>`
+export const PropGroupTag = styled.span<Parameters<typeof tagVariants>[0]>`
+  ${tagVariants}
   user-select: none;
   position: relative;
   display: inline-block;
-  padding: ${spacing[4]} ${spacing[8]};
+  padding: ${themed('spacing.4')} ${themed('spacing.8')};
   font-size: 11px;
-  font-family: ${({ theme }) => theme.typography.fonts.mono};
-  font-weight: 700;
+  font-family: ${themed('fontFamily.accent')};
   border-radius: 4px;
-  text-transform: uppercase;
   cursor: help;
-
-  &:first-of-type {
-    margin-left: 0;
-  }
+  border: 1px solid;
 
   &:hover {
     > * {
       display: block;
     }
   }
-
-  ${({ active, theme }) =>
-    active
-      ? css`
-          color: ${theme.color.secondary};
-          border: 1px solid ${theme.color.secondary};
-        `
-      : css`
-          color: ${theme.color.dark};
-          border: 1px solid ${theme.color.mediumdark};
-        `}
-`;
-
-export const Title = styled.p`
-  display: inline-block;
-  font-weight: 700;
-  margin: 0;
-`;
-
-export const ToggleContainer = styled.div`
-  display: flex;
-  align-items: flex-start;
 `;
 
 export const ToggleLabel = styled.label`
   margin-right: 4px;
-  font-size: 14px;
+  font-size: 16px;
   line-height: 2;
 `;
