+++ packages/styleguide/.storybook/components/Scales/ColorScale.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
-import { boxShadows } from '@codecademy/gamut-styles';
-import { css, styled } from '@storybook/theming';
-import { Box } from '../Box';
+import { boxShadows, themed } from '@codecademy/gamut-styles';
+import styled from '@emotion/styled';
+import { Box } from '@codecademy/gamut/src';
 
-const HexCode = styled.span(
-  (props) => css`
-    font-family: ${props.theme.fontCode};
-  `
-);
+const HexCode = styled.span`
+  font-family: ${themed('fontFamily.monospace')};
+  font-size: 12px;
+`;
+
+const Color = styled.div<{ bg?: string }>(({ bg }) => ({
+  backgroundColor: bg,
+}));
 
 export const ColorScale: React.FC<{ colors: Record<string, string> }> = ({
   colors,
@@ -26,19 +29,18 @@ export const ColorScale: React.FC<{ colors: Record<string, string> }> = ({
         boxShadow={boxShadows[1]}
       >
         {weights.map(([key, hex]) => (
-          <Box key={`color-${key}`} backgroundColor={hex} />
+          <Color key={`color-${key}`} bg={hex} />
         ))}
       </Box>
-      <Box width="100%" padding="4px" display="grid" gridAutoFlow="column">
+      <Box width="100%" padding={4} display="grid" gridAutoFlow="column">
         {weights.map(([key, hex]) => (
           <Box
             key={`hex-${key}`}
-            fontSize="12px"
             display="grid"
             alignItems="center"
             justifyContent="center"
             textAlign="center"
-            textColor="rgba(51,51,51,0.5)"
+            textColor="gray-400"
           >
             {weights.length > 1 && <HexCode>{key}</HexCode>}
             <HexCode>{hex}</HexCode>
