+++ packages/styleguide/.storybook/components/index.tsx
@@ -1,13 +1,11 @@
-export * from './Badge';
-export * from './Swatches';
-export * from './TextExample';
-export * from './Box';
-export * from './Scale';
+export * from './Markdown/Badge';
+export * from './Scales/Swatches';
+export * from './Scales/TextExample';
+export * from './Scales/Scale';
 export * from './PropsTable';
 export * from './DataTable';
-export * from './MarkdownElements';
-export * from './Box';
-export * from './ColorScale';
+export * from './Markdown/MarkdownElements';
+export * from './Scales/ColorScale';
 export * from './Page';
 export * from './TableOfContents';
 export * from './utils';
