+++ packages/styleguide/tsconfig.json
@@ -1,14 +1,19 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "/",
-    "baseUrl": ".storybook",
+    "baseUrl": ".",
     "paths": {
-      "~styleguide/blocks": ["components"]
+      "~styleguide/blocks": ["./.storybook/components"]
     },
     "allowJs": true,
     "noEmit": true,
     "declaration": false
   },
-  "include": ["../../typings/*.d.ts", "../**/src/**/*", "./**/*"]
+  "include": [
+    "../../typings/*.d.ts",
+    "./.storybook/**/*.ts",
+    "./.storybook/**/*.tsx",
+    "./stories.**/*.tsx",
+    "./stories.**/*.ts"
+  ]
 }
